Issue 319, SQL error on large domain
git-svn-id: svn://hmailserver.com:13600@74 b719dfc6-fa5d-b644-ba71-0e501e651431

+++ trunk/source/Server/Common/Persistence/PersistentDomain.h
@@ -25,7 +25,7 @@ namespace HM
       static bool DomainExists(const String &DomainName, bool &bIsActive);
 
       static int GetSize(shared_ptr<Domain> pDomain);
-      static int GetAllocatedSize(shared_ptr<Domain> pDomain);
+      static __int64 GetAllocatedSize(shared_ptr<Domain> pDomain);
 
    private:
       static bool _RenameDomain(shared_ptr<Domain> pDomain);
